Worked on Game Mode
Did some work on the game mode class and the team death match class. These two will work together to create an interchangeable game mode
