--- Assets/_Xandoo/Prefabs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 62d77801462c53c45812e44a1852ff79
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
