+++ Assets/_Xandoo/Scripts/SOBJ_TDMSettings_X.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: deff02e7725ac0c4f80cab1afd515d60
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
