+++ Assets/_Xandoo/Scripts/S_GameMode_X.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 34f362b2bc6a981439abf48b9128d577
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
