+++ Assets/_Xandoo/Scripts/S_TeamDeathMatch_X.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8622e09252c5cdd46ae16987f07e6e1b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
